requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.1
xgboost==2.0.0
joblib==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.21
alembic==1.12.1
redis==5.0.1

# API & Web
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
flask-socketio==5.3.4
python-socketio==5.9.0

# Monitoring & Logging
python-json-logger==2.0.7
structlog==23.2.0

# Data Processing
pytz==2023.3
python-dateutil==2.8.2

# ML & Stats
scipy==1.11.3
statsmodels==0.14.0

# Security
cryptography==41.0.7
PyJWT==2.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
unittest-mock==1.5.0

# Utilities
PyYAML==6.0.1
pydantic==2.4.2
